% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_directory}
\alias{cache_directory}
\alias{cache_info}
\alias{cache}
\title{On-disk cache management}
\usage{
cache_directory(id = census_id())

cache_info(id = census_id())

cache(id = census_id())
}
\arguments{
\item{id}{character(1) unique identifier associated with a
particular census snapshot, as returned by \code{census_id()}.}
}
\value{
\code{cache_directory()} returns a character(1) file path to the
location of the cache associated with the census with \code{id}.

\code{cache_info()} returns a tibble with file name, size, and
'mtime' (last-modified time). The mtime is used by
\code{cache_disk()} to manage the size and age of objects in the
cache.

\code{cache()} returns an object created by
\code{cachem::chach_disk()} that can be used to query and delete
items in the cache.
}
\description{
\code{cache_directory()} reports the path to the on-disk
cache for a particular census.

\code{cache_info()} summarizes file size and last
modification time of files in the cache.

\code{cache()} returns an object used to manage the cache.
}
\details{
The on-disk cache is created and managed by the
\code{cache_disk()} function in the cachem package. Consult that
package for details about working with this object.

File names in the cache are a hash of the function
arguments and body of which they are a cache; it is not
possible to know transparently which file corresponds to which
memoized function.
}
\examples{
cache_directory()

cache_info()

cache()

}
