% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census}
\alias{census}
\alias{census_id}
\alias{census_names}
\alias{census_versions}
\alias{show_package_versions}
\title{Obtain a reference to a CELLxGENE 'SOMA' collection}
\usage{
census(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

census_id(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

census_names(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

census_versions()
}
\arguments{
\item{version}{The version (date) of the census to use. \code{version = "stable"} indicates the most recent stable release; \code{version = "latest"} is the most recent release. Additional dates are
available with \code{census_versions()}.}

\item{uri}{The uri corresponding to census \code{version}; this is
usually discovered automatically.}

\item{tiledbsoma_ctx}{A 'context' providing mostly low-level flags
influencing the performance of tiledbsoma. One illustration of
this functionality is in the body of the \code{census()} function.}
}
\value{
\code{census()} returns a \code{tiledbsoma::SOMACollection} object. Details
of the census are available with \code{census()$get_metadata()}

\code{census_id()} returns the 7-character git commit sha that
uniquely identifies the current release of the census.

\code{census_names()} returns a character vector of possible
values. Use these values in calls to, e.g., \code{feature_data()} or
\code{observation_data()}.

\code{census_versions()} returns a tibble with a column of
available versions (release dates), and a 'status' column
indicating the 'stable' and 'latest' versions.
}
\description{
\code{census()} queries CELLxGENE for a particular census.

\code{census_id()} reports a unique identifier for a
particular census version.

\code{census_names()} queries the census for available
'experiments'. In CELLxGENE, experiments correspond to
organisms, e.g., \code{"homo_sapiens"} or \code{"mus_musculus"}.

\code{census_versions()} queries the CELLxGENE server for
available versions. Versions are denoted by release date.
}
\details{
\code{census()} is 'memoised', requiring high-latency internet access
only on its first use.
}
\examples{
census()
census()$get_metadata()

census_id()

census_names()

census_versions()

tiledbsoma::show_package_versions()
}
