% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\name{feature_data}
\alias{feature_data}
\alias{observation_data}
\alias{assay_data}
\title{Discover census features, observations, and assays (counts)}
\usage{
feature_data(organism, version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

observation_data(
  organism,
  version = "stable",
  uri = NULL,
  tiledbsoma_ctx = NULL
)

assay_data(
  organism,
  features,
  observations,
  version = "stable",
  uri = NULL,
  tiledbsoma_ctx = NULL
)
}
\arguments{
\item{organism}{one of the values returned by \code{census_names()},
specifically 'homo_sapiens' or 'mus_musculus' at the time of
writing this documentation.}

\item{version}{The version (date) of the census to use. \code{version = "stable"} indicates the most recent stable release; \code{version = "latest"} is the most recent release. Additional dates are
available with \code{census_versions()}.}

\item{uri}{The uri corresponding to census \code{version}; this is
usually discovered automatically.}

\item{tiledbsoma_ctx}{A 'context' providing mostly low-level flags
influencing the performance of tiledbsoma. One illustration of
this functionality is in the body of the \code{census()} function.}

\item{features}{a \code{tibble}, typically derived from
\code{feature_data()} via \code{filter()}, \code{select()}, etc., and
containing the \code{soma_joinid} column.}

\item{observations}{a \code{tibble}, typically derived from
\code{observation_data()} via \code{filter()}, \code{select()}, etc., and
containing the \code{soma_joinid} column.}
}
\value{
\code{feature_data()} returns a tibble with columns describing
each feature (gene) in \code{organism}.

\code{observation_data()} returns a dbplyr-based tibble of cell
annotations. An aesthetic problem is that the 'connection' to
the database is not available to the user, and duckdb warns
that \verb{Database is garbage-collected...}; this message can be
ignored.

\code{assay_data()} returns a sparse matrix (\code{dgCMatrix})
summarizing counts found in the census for the \code{soma_joinid}
columns of the \code{features} and \code{observations} tibble
arguments. The counts are from the 'raw' layer' of 'X'
collection of the 'RNA' measurement in the experiment.
}
\description{
\code{feature_data()} reports information about features
(genes) present in the census.

\code{observation_data()} reports information about all
cells in the census.

\code{assay_data()} queries the census for 'raw' counts for
the RNA-seq data corresponding to selected features and columns
in a census.
}
\details{
\code{observation_data()} is memoised to disk. The data is large (e.g.,
more than 50 million rows for \emph{Homo sapiens}) so the initial
download can be time-consuming (10's of minutes). During download
in an interactive session, the number of 'chunks' and records are
displayed; for the 2023-05-15 census of \code{homo_sapiens}, there were
more than 52 million records (cells) downloaded in 124 chunks.

The data are stored in a 'duckdb' database. The return value can be
used via \code{dbplyr} for very fast and memory efficient filtering,
selection, and summary.

Currently, for \code{assay_data()} and
\code{single_cell_experiment()}, the user must ensure that the
features, observations, and assay data are for the same census
and organism. Duplicate rows in \code{features} and \code{observations}
are allowed.
}
\examples{
feature_data("mus_musculus")

mus <- observation_data("mus_musculus")
mus

mus |>
    count(assay, sort = TRUE)

mus |>
    filter(grepl("diabetes", disease)) |>
    count(disease, sex, tissue)

## use features and observations as filters for assay (count) data
features <-
   feature_data("mus_musculus") |>
   ## rows 4, 3, 4 of the tibble
   slice(c(4:3, 4))

observations <-
   observation_data("mus_musculus") |>
   ## first two rows of the tibble
   head(2) |>
   collect()

assay_data("mus_musculus", features, observations)

}
