% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_info.R
\name{datasets}
\alias{datasets}
\alias{summary_cell_counts}
\alias{feature_data}
\title{Discover census datasets, cells, features, and observations}
\usage{
datasets(census = NULL)

summary_cell_counts(census = NULL)

feature_data(organism = c("homo_sapiens", "mus_musculus"), census = NULL)
}
\arguments{
\item{census}{a \code{tiledbsoma::SOMACollection} object as returned by
\code{census()}. If \code{NULL}, then the default returned by \code{census()}.}

\item{organism}{one of 'homo_sapiens' or 'mus_musculus'. Default:
'homo_sapiens'.}
}
\value{
\code{datasets()} returns a tibble with information about the
collections and datasets represented in the census.

\code{summary_cell_counts()} returns a tibble summarizing the
organism, facets (\code{category}, e.g., 'sex') and levels (\code{label},
e.g., 'female') represented in the data, and unique and total
cell counts in each facet and level.

\code{feature_data()} returns a tibble with columns describing
each feature (gene) in \code{organism}.
}
\description{
\code{datasets()} queries CELLxGENE for datasets used in
constructing the census.

\code{summary_cell_counts()} reports the facets (e.g., sex)
and levels (e.g., male, female) in the census, and the number
of cells associated with each facet and level.

\code{feature_data()} reports information about features
(genes) present in the census.
}
\details{
\code{datasets()}, \code{summary_cell_counts()}, and \code{feature_data()} are
'memoised' so that they are only expensive on their first use. The
'tibble' returned by these functions is memoised to disk, so that
re-using the function is fast even across sessions. See
\code{?cache_info()} for details on cache management.
}
\examples{
datasets() |>
    glimpse()

summary_cell_counts() |>
    count(category)

## number of cells from female, male, and 'unknown' samples in
## humans and mice
summary_cell_counts() |>
    filter(category == "sex") |>
    select(
        organism, label,
        unique_cell_count, total_cell_count
    )

feature_data()  # default: homo_sapiens
feature_data("mus_musculus")

}
