% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleCellExperiment.R
\name{single_cell_experiment}
\alias{single_cell_experiment}
\title{Construct a SingleCellExperiment from census data}
\usage{
single_cell_experiment(
  organism,
  features,
  observations,
  version = "stable",
  uri = NULL,
  tiledbsoma_ctx = NULL
)
}
\arguments{
\item{organism}{one of the values returned by \code{census_names()},
specifically 'homo_sapiens' or 'mus_musculus' at the time of
writing this documentation.}

\item{features}{a \code{tibble}, typically derived from
\code{feature_data()} via \code{filter()}, \code{select()}, etc., and
containing the \code{soma_joinid} column.}

\item{observations}{a \code{tibble}, typically derived from
\code{observation_data()} via \code{filter()}, \code{select()}, etc., and
containing the \code{soma_joinid} column.}

\item{version}{The version (date) of the census to use. \code{version = "stable"} indicates the most recent stable release; \code{version = "latest"} is the most recent release. Additional dates are
available with \code{census_versions()}.}

\item{uri}{The uri corresponding to census \code{version}; this is
usually discovered automatically.}

\item{tiledbsoma_ctx}{A 'context' providing mostly low-level flags
influencing the performance of tiledbsoma. One illustration of
this functionality is in the body of the \code{census()} function.}
}
\description{
\code{single_cell_experiment()} queries the census for
assay data corresponding to features and observations, and
assembles the result into a SingleCellExperiment. The count
data are accessible using \code{SingleCellExperiment::counts()}.
}
\details{
\code{single_cell_experiment()} requires that the
SingleCellExperiment Bioconductor package is installed, e.g.,
via \code{BiocManager::install("SingleCellExperiment")}.
}
\examples{
features <-
   feature_data("mus_musculus") |>
   ## rows 4, 3, 4 of the tibble
   slice(c(4:3, 4))

observations <-
   observation_data("mus_musculus") |>
   ## first two rows of the tibble
   head(2) |>
   collect()

single_cell_experiment("mus_musculus", features, observations)

}
