% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{census}
\alias{census}
\alias{census_id}
\alias{census_names}
\title{Obtain a reference to a CELLxGENE 'SOMA' collection}
\usage{
census(census_version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

census_id(...)

census_names(...)
}
\arguments{
\item{census_version}{see \code{?cellxgene.census::open_soma}}

\item{uri}{see \code{?cellxgene.census::open_soma}}

\item{tiledbsoma_ctx}{see \code{?cellxgene.census::open_soma}}

\item{...}{arguments passed to \code{census()}.}

\item{census}{an object returned by \code{census()}; if present, this
overrides specifications in \code{...}.}
}
\value{
\code{census()} returns a \code{tiledbsoma::SOMACollection} object. Details
of the \code{census_version} are available with
\code{census()$get_metadata()}

\code{census_id()} returns the 7-character git commit sha that
uniquely identifies the current release of the census.

\code{census_names()} returns a character vector of possible
values. Use these values in calls to, e.g., \code{feature_data()} or
\code{observation_data()}.
}
\description{
\code{census()} queries CELLxGENE for a particulalr census.

\code{census_id()} reports a unique identifier for a
particular census version.

\code{census_names()} queries the census for available
'experiments'. In CELLxGENE, experiments correspond to
organisms, e.g., \code{"homo_sapiens"} or \code{"mus_musculus"}.
}
\details{
\code{census()} is 'memoised', requiring high-latency internet access
only on its first use.
}
\examples{
census()
census()$get_metadata()

census_id()

census_names()

}
