% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_info.R
\name{census}
\alias{census}
\alias{datasets}
\alias{summary_cell_counts}
\alias{feature_data}
\title{Discover metadata about CELLxGENE census}
\usage{
census(census_version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

datasets(census = NULL)

summary_cell_counts(census = NULL)

feature_data(organism = c("homo_sapiens", "mus_musculus"), census = NULL)
}
\arguments{
\item{census_version}{see \code{?cellxgene.census::open_soma}}

\item{uri}{see \code{?cellxgene.census::open_soma}}

\item{tiledbsoma_ctx}{see \code{?cellxgene.census::open_soma}}

\item{census}{a \code{tiledbsoma::SOMACollection} object as returned by
\code{census()}. If \code{NULL}, then the default returned by \code{census()}.}

\item{organism}{one of 'homo_sapiens' or 'mus_musculus'. Default:
'homo_sapiens'.}
}
\value{
\code{census()} returns a \code{tiledbsoma::SOMACollection} object. Details
of the \code{census_version} are available with
\code{census()$get_metadata()}

\code{datasets()} returns a tibble with information about the
collections and datasets represented in the census.

\code{summary_cell_counts()} returns a tibble summarizing the organism,
factors (\code{category}) and levels represented in the data, and unique
and total cell counts in each category.

\code{feature_data()} returns a tibble with columns describing each
feature (gene) in \code{organism}.
}
\description{
Discover metadata about CELLxGENE census
}
\details{
\code{census()} is 'memoised', requiring high-latency internet access
only on its first use.

\code{datasets()}, \code{summary_cell_counts()}, and \code{feature_data()} are
'memoised' so that they are only expensive on their first use. The
'tibble' returned by these functions is memoised to disk, so that
re-using the function is fast even across sessions. See
\code{?cache_info()} for details on cache management.
}
\examples{
census()
census()$get_metadata()

datasets() |>
    glimpse()

summary_cell_counts() |>
    count(category)

## number of cells from female, male, and 'unknown' samples in
## humans and mice
summary_cell_counts() |>
    filter(category == "sex") |>
    select(
        organism, label,
        unique_cell_count, total_cell_count
    )

feature_data()  # default: homo_sapiens
feature_data("mus_musculus")

}
