% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_info.R
\name{datasets}
\alias{datasets}
\alias{summary_cell_counts}
\title{Discover census datasets and cell count summaries}
\usage{
datasets(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)

summary_cell_counts(version = "stable", uri = NULL, tiledbsoma_ctx = NULL)
}
\arguments{
\item{version}{The version (date) of the census to use. \code{version = "stable"} indicates the most recent stable release; \code{version = "latest"} is the most recent release. Additional dates are
available with \code{census_versions()}.}

\item{uri}{The uri corresponding to census \code{version}; this is
usually discovered automatically.}

\item{tiledbsoma_ctx}{A 'context' providing mostly low-level flags
influencing the performance of tiledbsoma. One illustration of
this functionality is in the body of the \code{census()} function.}
}
\value{
\code{datasets()} returns a tibble with information about the
collections and datasets represented in the census.

\code{summary_cell_counts()} returns a tibble summarizing the
organism, facets (\code{category}, e.g., 'sex') and levels (\code{label},
e.g., 'female') represented in the data, and unique and total
cell counts in each facet and level.
}
\description{
\code{datasets()} queries CELLxGENE for datasets used in
constructing the census.

\code{summary_cell_counts()} reports the facets (e.g., sex)
and levels (e.g., male, female) in the census, and the number
of cells associated with each facet and level.
}
\details{
\code{datasets()}, \code{summary_cell_counts()}, an \code{feature_data()} are
'memoised' so that they are only expensive on their first use. The
'tibble' returned by these functions is memoised to disk, so that
re-using the function is fast even across sessions. See
\code{?cache_info()} for details on cache management.
}
\examples{
datasets() |>
    glimpse()

summary_cell_counts() |>
    count(category)

## number of cells from female, male, and 'unknown' samples in
## humans and mice
summary_cell_counts() |>
    filter(category == "sex") |>
    select(
        organism, label,
        unique_cell_count, total_cell_count
    )

}
